# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goat Farm Management API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/animals": {
      "get": {
        "tags": [
          "animals"
        ],
        "summary": "List Animals",
        "operationId": "list_animals_api_animals_get",
        "parameters": [
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "QUARANTINE",
                    "FOUNDATION",
                    "BREEDING",
                    "PREGNANCY_EARLY",
                    "PREGNANCY_LATE",
                    "DELIVERY",
                    "RECOVERY",
                    "RESTING",
                    "MALE_KIDS",
                    "FEMALE_KIDS"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bucket"
            }
          },
          {
            "name": "sex",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "M",
                    "F"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sex"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ACTIVE",
                    "SOLD",
                    "DEAD",
                    "CULLED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "animals"
        ],
        "summary": "Create Animal",
        "operationId": "create_animal_api_animals_post",
        "parameters": [
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/animals/{animal_id}": {
      "get": {
        "tags": [
          "animals"
        ],
        "summary": "Animal Profile",
        "operationId": "animal_profile_api_animals__animal_id__get",
        "parameters": [
          {
            "name": "animal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Animal Id"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/animals/{animal_id}/move": {
      "post": {
        "tags": [
          "animals"
        ],
        "summary": "Move Bucket",
        "operationId": "move_bucket_api_animals__animal_id__move_post",
        "parameters": [
          {
            "name": "animal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Animal Id"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/animals/{animal_id}/weight": {
      "post": {
        "tags": [
          "animals"
        ],
        "summary": "Record Weight",
        "operationId": "record_weight_api_animals__animal_id__weight_post",
        "parameters": [
          {
            "name": "animal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Animal Id"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightRecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/animals/{animal_id}/status": {
      "post": {
        "tags": [
          "animals"
        ],
        "summary": "Change Status",
        "operationId": "change_status_api_animals__animal_id__status_post",
        "parameters": [
          {
            "name": "animal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Animal Id"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChangeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnimalCreateIn": {
        "properties": {
          "tag_number": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Tag Number"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sex": {
            "type": "string",
            "enum": [
              "M",
              "F"
            ],
            "title": "Sex"
          },
          "source": {
            "type": "string",
            "enum": [
              "BORN",
              "PURCHASED"
            ],
            "title": "Source"
          },
          "current_bucket": {
            "type": "string",
            "enum": [
              "QUARANTINE",
              "FOUNDATION",
              "BREEDING",
              "PREGNANCY_EARLY",
              "PREGNANCY_LATE",
              "DELIVERY",
              "RECOVERY",
              "RESTING",
              "MALE_KIDS",
              "FEMALE_KIDS"
            ],
            "title": "Current Bucket"
          },
          "breed": {
            "type": "string",
            "maxLength": 60,
            "title": "Breed",
            "default": "Osmanabadi"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "estimated_dob": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Dob"
          },
          "birth_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "SINGLE",
                  "TWIN",
                  "TRIPLET",
                  "QUADRUPLET",
                  "MULTIPLET"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Type"
          },
          "birth_weight": {
            "anyOf": [
              {
                "type": "number",
                "le": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Weight"
          },
          "purchase_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Date"
          },
          "purchase_price": {
            "anyOf": [
              {
                "type": "number",
                "le": 1000000000
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Price"
          },
          "seller_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Name"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "le": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "tag_number",
          "sex",
          "source",
          "current_bucket"
        ],
        "title": "AnimalCreateIn"
      },
      "AnimalListOut": {
        "properties": {
          "animals": {
            "items": {
              "$ref": "#/components/schemas/AnimalOut"
            },
            "type": "array",
            "title": "Animals"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "animals",
          "total"
        ],
        "title": "AnimalListOut"
      },
...
      "AnimalProfileOut": {
        "properties": {
          "animal": {
            "$ref": "#/components/schemas/AnimalOut"
          },
          "kids": {
            "items": {
              "$ref": "#/components/schemas/AnimalOut"
            },
            "type": "array",
            "title": "Kids"
          },
          "weights": {
            "items": {
              "$ref": "#/components/schemas/WeightRecordOut"
            },
            "type": "array",
            "title": "Weights"
          },
          "moves": {
            "items": {
              "$ref": "#/components/schemas/BucketMoveOut"
            },
            "type": "array",
            "title": "Moves"
          },
          "health_events": {
            "items": {
              "$ref": "#/components/schemas/HealthEventOut"
            },
            "type": "array",
            "title": "Health Events"
          },
          "breedings": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Breedings"
          }
        },
        "type": "object",
        "required": [
          "animal",
          "kids",
          "weights",
          "moves",
          "health_events",
          "breedings"
        ],
        "title": "AnimalProfileOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoveIn": {
        "properties": {
          "to_bucket": {
            "type": "string",
            "enum": [
              "QUARANTINE",
              "FOUNDATION",
              "BREEDING",
              "PREGNANCY_EARLY",
              "PREGNANCY_LATE",
              "DELIVERY",
              "RECOVERY",
              "RESTING",
              "MALE_KIDS",
              "FEMALE_KIDS"
            ],
            "title": "To Bucket"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "to_bucket"
        ],
        "title": "MoveIn"
      },
...
      "WeightIn": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg",
            "le": 1000
          },
          "bcs": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bcs"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "weight_kg"
        ],
        "title": "WeightIn"
      },
...
    }
  }
}